refactor(loadpoint) + feat(ev): poke + bubble label cleanup
Three small things in one commit because they touched the same
plug-in flow:

1. Refactor: pulled vehicle-discovery + auto-charge helpers into
   their own files. loadpoint/loadpoint.go was creeping past
   600 lines with three separate concerns interleaved (Manager
   core, plug-in scoring, schedule auto-posting). New layout:
     - loadpoint.go         — Config + Policy + State + Manager
     - vehicle_discovery.go — VehicleSample/{Telemetry,Snapshot,
                              Poke}Func types + pickPluggedVehicle
     - auto_charge.go       — applyAutoSchedule + HH:MM helpers
   No behaviour change; just file moves.

2. Plug-in poke: drivers/registry.go gains PokePoll(name) which
   queues a "poll" command type onto the driver's cmdCh. The
   runLoop's command handler invokes driver.Poll out-of-cycle and
   updates DriverHealth like a regular tick. Manager.Observe now
   pokes every known vehicle driver on EV plug-in transition so
   the auto-discovery scoring sees fresh BMS readings instead of
   up-to-60-s-stale ones. Best-effort + non-blocking; the channel
   has 8-slot buffer and skips poke when full (next scheduled tick
   covers the gap).

3. EV bubble label: the planet sub-text now sources the charger
   state-label from the driver itself (Easee's ev_state_label
   etc.) instead of binarising on "charging vs not". Critically,
   when ev_connected is false the bubble reads "disconnected"
   instead of the ambiguous "idle" — matches what the per-driver
   Advanced card shows for the same charger.
